Refactor + fix "off" method
